requirements.txt
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.4
boto3==1.40.60
botocore==1.40.60
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.5
defusedxml==0.7.1
Django==5.2.3
django-background-tasks==1.2.8
django-cors-headers==4.7.0
django-filter==25.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.3
dotenv==0.9.9
drf-yasg==1.21.10
gunicorn==21.2.0
h11==0.16.0
hdrezka==4.0.2
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jmespath==1.0.1
load-dotenv==0.1.0
lxml==5.4.0
oauthlib==3.3.1
packaging==25.0
pillow==11.2.1
pycparser==2.22
PyJWT==2.9.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
pyzbar==0.1.9
qrcode==8.2
reportlab==4.4.2
requests==2.32.4
requests-oauthlib==2.0.0
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
youtube-dl==2021.12.17
